Create Staff Module
