+++ src/systems.cpp
@@ -45,44 +45,3 @@ void systemUpdateSimDate(GameResource &game) {
   game.day++;
   spdlog::info("It's Day {}", game.day);
 }
-
-void company_generator_system(flecs::iter &it) {
-  spdlog::debug("Creating teams and employees");
-  const flecs::world &world = it.world();
-
-  auto finance = world.entity("Finance").set<Team>({"Finance"});
-  auto purchasing = world.entity("Purchasing")
-                        .set<Team>({"Purchasing"})
-                        .add<TeamMember>(finance);
-  auto payable = world.entity("Payable")
-                     .set<Team>({"Accounts Payable"})
-                     .add<TeamMember>(finance);
-
-  world.entity("Anna")
-      .set<Person>({"Anna", "Delaney"})
-      .add<Employee>()
-      .add<TeamMember>(purchasing);
-  world.entity("Bernardo")
-      .set<Person>({"Bernardo", "Jimenez"})
-      .add<Employee>()
-      .add<TeamMember>(purchasing);
-  world.entity("Kathy")
-      .set<Person>({"Kathy", "Lau"})
-      .add<Employee>()
-      .add<TeamMember>(purchasing)
-      .add<Manager>(purchasing);
-  world.entity("Dave")
-      .set<Person>({"Dave", "Jones"})
-      .add<Employee>()
-      .add<TeamMember>(finance)
-      .add<Manager>(finance);
-  world.entity("Elena")
-      .set<Person>({"Elena", "Rodriguez"})
-      .add<Employee>()
-      .add<TeamMember>(payable)
-      .add<Manager>(payable);
-  world.entity("Freddie")
-      .set<Person>({"Fred", "Dorn"})
-      .add<Employee>()
-      .add<TeamMember>(payable);
-}
